+++ scarletio/http_server/idea.py
@@ -186,7 +186,7 @@ class HTTPRequestHandler(HttpReadWriteProtocol):
         """
         HttpReadWriteProtocol.connection_made(self, transport)
         self.server.register(self)
-        self.handler_task = Task(self.lifetime_handler(), self._loop)
+        self.handler_task = Task(self._loop, self.lifetime_handler())
     
     async def lifetime_handler(self):
         try:
@@ -392,7 +392,7 @@ class HTTPServer:
         """
         close_connection_task = self.close_connection_task
         if close_connection_task is None:
-            close_connection_task = Task(self._close(), self.loop)
+            close_connection_task = Task(self.loop, self._close())
             self.close_connection_task = close_connection_task
         
         return close_connection_task